+++ TestApps/SIL.Windows.Forms.TestApp/ScrReferenceFilterDlg.Designer.cs
@@ -1,14 +1,3 @@
-// ---------------------------------------------------------------------------------------------
-#region // Copyright (c) 2022, SIL International.
-// <copyright from='2011' to='2022' company='SIL International'>
-//		Copyright (c) 2022, SIL International.
-//
-//		Distributable under the terms of the MIT License (http://sil.mit-license.org/)
-// </copyright>
-#endregion
-//
-// File: ScrReferenceFilterDlg.Designer.cs
-// ---------------------------------------------------------------------------------------------
 namespace SIL.Windows.Forms.TestApp
 {
 	partial class ScrReferenceFilterDlg